`@remotion/studio-codemods`: Split JSX without Prettier
