+++ packages/studio-server/src/preview-server/routes/insert-jsx-element.ts
@@ -18,7 +18,6 @@ import {
 	pushToUndoStack,
 	suppressUndoStackInvalidation,
 } from '../undo-stack';
-import {warnAboutPrettierOnce} from './log-updates/log-update';
 import {
 	getCodemodTimingPrefix,
 	withSourceFileWriteQueue,
@@ -248,7 +247,6 @@ export const insertJsxElementHandler: ApiHandler<
 				source,
 				oldContents,
 				output,
-				formatted,
 				insertedNodePath,
 				logLine,
 				nodePathRemappings,
@@ -305,13 +303,9 @@ export const insertJsxElementHandler: ApiHandler<
 				{indent: false, logLevel},
 				`${getCodemodTimingPrefix(logLevel)}${RenderInternals.chalk.blueBright(`${locationLabel}`)} Added ${elementLabel}`,
 			);
-			if (!formatted) {
-				warnAboutPrettierOnce(logLevel);
-			}
-
 			RenderInternals.Log.verbose(
 				{indent: false, logLevel},
-				`[insert-jsx-element] Wrote ${source}${formatted ? ' (formatted)' : ''}`,
+				`[insert-jsx-element] Wrote ${source}`,
 			);
 
 			printUndoHint(logLevel);
